fix the folder rename
